Move external modules to separate subdirectory.
