+++ coxeter/extern/bentley_ottmann/LICENSE
