+++ coxeter/extern/bentley_ottmann/poly_point_isect.py
