+++ coxeter/extern/polytri/LICENSE
