+++ coxeter/extern/polytri/polytri.py
