+++ coxeter/shapes/polygon.py
@@ -5,8 +5,8 @@ import warnings
 import numpy as np
 import rowan
 
-from ..bentley_ottmann import poly_point_isect
-from ..polytri import polytri
+from ..extern.bentley_ottmann import poly_point_isect
+from ..extern.polytri import polytri
 from .base_classes import Shape2D
 from .circle import Circle
 from .utils import _generate_ax, rotate_order2_tensor, translate_inertia_tensor
