+++ setup.cfg
@@ -15,7 +15,7 @@ python-tag = py3
 
 [flake8]
 max-line-length = 88
-exclude = coxeter/polytri/*,coxeter/bentley_ottmann/*
+exclude = coxeter/extern/*
 # Use select to ignore unwanted flake8 plugins
 select = E,F,W,BLK,B,D,RST,N,I
 per-file-ignores = 
@@ -31,7 +31,7 @@ rst-roles =
 	attr,class,func,meth,mod,obj,ref,term,cite
 
 [isort]
-skip_glob = */bentley_ottmann/*,*/polytri/*
+skip_glob = */extern/*
 multi_line_output = 3
 include_trailing_comma = True
 force_grid_wrap = 0
